do not do further validation on cells that already have datatype violations
We skip further validation on such cells since this has the potential
(in a strongly typed database) to result in database errors. For
example, suppose we are trying to validate the value in a cell that is
of text type when it is supposed to be of integer type (or vice versa).
When we then construct the SQL query to compare that cell with the
values of other cells in the same column, the database will complain.