do not unnecessarily strip _conflict from table names.
This is not necessary in most cases and makes the code confusing.
There was a need, previously, because we used to call the
update/insert/delete single row functions directly on the conflict
table in some cases - the reason being that the update_row function
had two possible paths, one direct and one indirect. Now that it is
always indirect we can clean this up.
The only time where it is still needed is when we are actually
generating insert statements.